fix: tests working again, validation problem fixed, refactor: removed unnecessary functions and added remaining specific status codes

+++ backend/src/validation/validation.ts
@@ -87,7 +87,7 @@ export const updateUserZodSchema = createInsertSchema(userSchema, {
 export const updateScoreZodSchema = z.object({
   streak: z.number(),
   longestStreak: z.number().optional(),
-  recentScores: z.array(z.number()),
+  recentScores: z.array(z.number().int())
 });
 
 export const updateDailyScoreZodSchema = z.object({
